Handle greased pushUps with large-arity childArrow
Previously there was a hidden assumption that the childArrow had arity 2
(e.g. &exist;). Now it should work for any arity (e.g &int;), though there may
still be a hidden assumption that the binding var to be greased is the first
arg.

As a test, adds a proof of sbor which fails without this change.