Support "fancy" distributes.
These have different operators in the consequent than the antecedent. The idea
is to allow using anchors inside a conjuction rather than just an implication.

I'm pretty certain that this cannot work without more changes, because the only
place that looks up distributes still uses the same argument twice. But
mysteriously, it seems to work already. Something must be wrong.

Also adds some more le goals to unlock it as a new root and bring us up to 171
thms proved.