Fix a pair of heinous engine bugs.
JS was helpfully casting an array to an int in a != statement. Also it was
helpfully converting an array to an int in an object lookup.

Also adds a new thm to land-le, whose proof happened to expose both bugs by
introducing a dummy variable for 0, which ended up mapping 2 -> [2].

Also adds the new goto url syntax.