Fix an engine dummy unification bug.
If a var was mapped to a dummy in the course of unifying an expression, and
later in that expression was mapped to a different dummy, this could overwrite
the earlier mapping rather than forcing the two dummies to collapse into one.

This is exercised by the in-progress proof of min-nex=0 (U5B3m) but it also
fails for other reasons so isn't being checked in yet.